delete mixin and move the logic to a new service instead
