+++ app/routes/project.js
@@ -1,13 +1,18 @@
 import Ember from 'ember';
 import semverCompare from 'npm:semver-compare';
 import _ from 'lodash/lodash';
-import ResetScrollPositionMixin from 'ember-api-docs/mixins/reset-scroll-position';
 
 
-export default Ember.Route.extend(ResetScrollPositionMixin, {
+const { inject } = Ember;
+
+export default Ember.Route.extend({
+
+  scrollPositionReset: inject.service(),
+
   model(params) {
     return this.store.find('project', params.project);
   },
+
   afterModel(project /*, transition */) {
     return project.get('projectVersions').then(versions => {
       const last = versions.toArray().sort((a, b) => {
@@ -17,5 +22,16 @@ export default Ember.Route.extend(ResetScrollPositionMixin, {
       })[versions.length - 1];
       return this.transitionTo('project-version', last);
     });
+  },
+
+  actions: {
+
+    willTransition(transition) {
+      this.get('scrollPositionReset').scheduleReset(transition);
+    },
+
+    didTransition() {
+      this.get('scrollPositionReset').doReset();
+    }
   }
 });
